invensis_pmc/requirements.txt
django==1.9.7
django-appconf==1.0.1
django-crispy-forms==1.6.0
django-extensions==1.6.7
django-filter==0.14.0
django-import-export==0.5.1
django-ipware==1.1.3
django-model-utils==2.3.1
django-mptt==0.8.4
Django-Select2==5.8.4
django-suit==0.2.17
future==0.15.2
gunicorn==19.6.0
html5lib==1.0b8
jsonpickle==0.9.3
oauthlib==1.0.3
Pillow==3.3.1
psycopg2==2.6.2
pycrypto==2.6.1
PyJWT==1.4.2
PyPDF2==1.26.0
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.21
reportlab==3.3.0
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
tablib==0.11.4
webencodings==0.5
xhtml2pdf==0.0.6
xlwt==1.2.0
